requirements.txt
altgraph==0.16.1
apipkg==1.5
asgiref==3.2.7
astroid==2.3.2
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.2.0
beautifulsoup4==4.7.1
binwalk==2.1.0
bs4==0.0.1
certifi==2018.11.29
cffi==1.14.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==5.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
Django==3.0.6
django-grappelli==2.14.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
enum34==1.1.6
execnet==1.7.1
Flask==1.0.2
future==0.18.2
html5lib==1.0.1
idna==2.8
imageio==2.9.0
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.16.0
keyboard==0.13.5
Kivy==1.11.1
kivy-deps.angle==0.1.9
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.2
lxml==4.3.3
MarkupSafe==1.1.0
matplotlib==3.3.0
mccabe==0.6.1
more-itertools==7.2.0
murl==0.5.1
networkx==2.4
nginx==0.0.1
numpy==1.17.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.1.26
packaging==19.1
pandas==1.0.5
pefile==2019.4.18
Pillow==6.0.0
pluggy==0.13.0
progressbar==2.5
PTable==0.9.2
py==1.8.0
PyAutoGUI==0.9.44
pycparser==2.20
pygame==1.9.6
PyGetWindow==0.0.5
Pygments==2.4.2
PyInstaller==3.5
PyJWT==1.7.1
pylint==2.4.3
PyMsgBox==1.0.6
pynput==1.6.8
pyparsing==2.4.2
pyperclip==1.7.0
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.15
pyqt5-tools==5.11.3.1.4
PyRect==0.1.4
pyscard==1.9.9
PyScreeze==0.1.21
pyTelegramBotAPI==3.6.6
pytest==5.1.2
pytest-cov==2.8.1
pytest-forked==1.1.3
pytest-xdist==1.31.0
python-dateutil==2.8.1
python-dotenv==0.10.1
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
pywin32==224
pywin32-ctypes==0.2.0
recordtype==1.3
requests==2.21.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.8
sqlparse==0.3.1
telebot==0.0.3
threadpoolctl==2.1.0
tifffile==2020.7.24
tqdm==4.31.1
typed-ast==1.4.0
typing==3.7.4.1
urllib3==1.24.1
urlopen==1.0.0
urlparse2==1.1.1
virtualenv==16.7.7
vk==2.0.2
vk-api==11.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wget==3.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0
